requirements.txt
aiohttp==3.8.1
gunicorn==20.1.0
uvloop==0.16.0
aiohttp-jinja2==1.5
aiohttp_cors==0.7.0
git+https://github.com/xoolo-eng/url_control.git@master
cryptography==37.0.2
aiosmtplib==1.1.6
aioredis==2.0.1
pyjwt==2.4.0